rfid_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
loguru==0.7.2
pydantic-settings==2.6.1
mysql-connector-python==9.1.0
